LayoutTests/fast/dom/shadow/get-element-by-id-in-shadow-root-expected.txt

Tests to ensure that internals.getElementByIdInShadowRoot can get an element in ShadowRoot by its id. Can only run within DRT

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getElementInShadowTreeStack('divA/inputB').id is "inputB"
PASS getElementInShadowTreeStack('divA/divC').id is "divC"
PASS getElementInShadowTreeStack('divA/divC/inputD').id is "inputD"
PASS getElementInShadowTreeStack('divA/divC//inputE').id is "inputE"
PASS getElementInShadowTreeStack('divA/divC///inputF').id is "inputF"
PASS successfullyParsed is true

TEST COMPLETE
